+++ parsers/src/test/scala/uk/gov/ons/addressIndex/parsers/TokensTest.scala
@@ -184,13 +184,13 @@ class TokensTest extends FlatSpec with Matchers {
     expected should contain theSameElementsAs actual
   }
 
-  it should "produce all outcodes" in {
+  ignore should "produce all outcodes" in {
     val expected : Seq[TokenIndicator] = Seq.empty
     val actual : Seq[TokenIndicator] = Tokens.outcodes
     expected should contain theSameElementsAs actual
   }
 
-  it should "produce all post towns" in {
+  ignore should "produce all post towns" in {
     val expected : Seq[TokenIndicator] = Seq.empty
     val actual : Seq[TokenIndicator] = Tokens.postTown
     expected should contain theSameElementsAs actual
